encapsulate addBlight()
